pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tdd-llm"
version = "0.1.0"
description = "Deploy TDD workflow templates for Claude and Gemini AI assistants"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Maxim", email = "maxim@example.com" }
]
keywords = ["tdd", "claude", "gemini", "ai", "workflow", "templates"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
tdd-llm = "tdd_llm.cli:app"

[project.urls]
Homepage = "https://github.com/user/tdd-llm"
Repository = "https://github.com/user/tdd-llm"

[tool.hatch.build.targets.wheel]
packages = ["src/tdd_llm"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
